service

+++ src/main/java/pl/edu/agh/to2/example/models/dto/WeatherRequestDto.java
@@ -1,4 +1,4 @@
 package pl.edu.agh.to2.example.models.dto;
 
-public record WeatherRequestDto(int lat, int lng) {
+public record WeatherRequestDto(double lat, double lng) {
 }
