+++ src/main/java/pl/edu/agh/to2/example/models/weather/WeatherRequest.java
@@ -1,4 +1,4 @@
 package pl.edu.agh.to2.example.models.weather;
 
-public record WeatherRequest(int lat, int lng) {
+public record WeatherRequest(double lat, double lng) {
 }
